Align BridgeState to streamline moving funds

+++ solidity/contracts/bridge/Bridge.sol
@@ -1207,7 +1207,7 @@ contract Bridge is Governable, EcdsaWalletOwner, Initializable {
         uint64 movedFundsSweepTxMaxTotalFee,
         uint32 movedFundsSweepTimeout,
         uint96 movedFundsSweepTimeoutSlashingAmount,
-        uint256 movedFundsSweepTimeoutNotifierRewardMultiplier
+        uint64 movedFundsSweepTimeoutNotifierRewardMultiplier
     ) external onlyGovernance {
         self.updateMovingFundsParameters(
             movingFundsTxMaxTotalFee,
